move get_or_launch_is to its own module
factored out of https://github.com/matrix-org/matrix-is-tester/pull/4
