+++ matrix_is_tester/launch_is.py
@@ -1,0 +1,56 @@
+#!/usr/bin/env python
+
+# -*- coding: utf-8 -*-
+
+# Copyright 2019 The Matrix.org Foundation C.I.C.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import atexit
+
+try:
+    from matrix_is_test.launcher import MatrixIsTestLauncher
+except ImportError:
+    print("ERROR: Couldn't import launcher")
+    print(
+        "matrix_is_tester needs an identity server to test: make sure, "
+        "'matrix_is_test.launcher.MatrixIsTestLauncher' is in "
+        "sys.path"
+    )
+
+    raise
+
+launcher = None
+base_url = None
+
+
+def get_or_launch_is():
+    global launcher
+    global base_url
+
+    if launcher is not None:
+        return base_url
+
+    launcher = MatrixIsTestLauncher()
+    base_url = launcher.launch()
+
+    atexit.register(destroy_is)
+
+    return base_url
+
+
+def destroy_is():
+    global launcher
+
+    if launcher is not None:
+        launcher.tearDown()
